Fix admin

+++ social media/src/main/java/com/codegym/socialmedia/model/social_action/FriendshipId.java
@@ -11,7 +11,7 @@ import java.io.Serializable;
 @AllArgsConstructor
 @Embeddable
 public class FriendshipId implements Serializable {
-    private Integer requesterId;
-    private Integer addresseeId;
+    private Long requesterId;
+    private Long addresseeId;
 
 }
